Combine test_enkf and test_enkf_main
test_enkf contained unit-tests for enkf_main and so
should be in test_enkf_main
